Add aubio-object abstraction
The idea behind it is to stop repeating the definition of a slot
`internal-something` for every object ported to the API.
Maybe a more generic `clean` function can arise from this refactor.

Also, having made all this changes, I realized that a lot of getters and setters
are literal translations of their C counterpart (shocking, right? a thing that
every FFI library already knew). I'm currently thinking in introducing a way to
write this kind of things as slots in the Lisp class and let it generate the
reader and writer methods for me