Add sink object
I found out reading Aubio source code that it isn't mandatory to keep the URI
string for sources and sink in memory, because Aubio itself copies it to keep
track of it. So I removed the code that keeps the URI C string in the `source`
object.

Added a way to ask the `float-matrix` it's total size (it was needed for the
`sink` example).

Added a `:source-variable` keyword argument to the `do-source` macro. I thought
I needed it for the `sink` example but it turn out not being necessary. I kept
it anyway because I think it's useful.